Rename data_validation.py -> input_mapping.py
